use rgba32 instead rgba8

+++ src/01-init/launch_params.h
@@ -2,7 +2,7 @@
 
 struct LaunchParams {
     int frame_id;
-    unsigned int* color_buffer;
+    float* color_buffer;
     int frame_width;
     int frame_height;
 };
